+++ eslint.config.mjs
@@ -243,7 +243,7 @@ export default defineConfigWithVueTs(
         {
           patterns: [
             {
-              group: ["electron"],
+              regexp: "^electron",
               message:
                 "このファイル内でelectronはimportできません。許可されているファイル内へ移すか、ESLintの設定を見直してください",
             },
